ipa: rpi: controller: awb: Separate Bayesian Awb into AwbBayes
Move parts of the AWB algorithm specific to the Bayesian algorithm into a
new class. This will make it easier to add new Awb algorithms in the future.
